+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/texture/TextureVideoPlayer.java
@@ -79,8 +79,13 @@ public final class TextureVideoPlayer extends VideoPlayer
   @NonNull
   @Override
   protected ExoPlayerEventListener createExoPlayerEventListener(@NonNull ExoPlayer exoPlayer) {
+    System.out.println(surfaceProducer);
+    boolean surfaceProducerHandlesCropAndRotation = surfaceProducer.handlesCropAndRotation();
     return new TextureExoPlayerEventListener(
-        exoPlayer, videoPlayerEvents, playerHasBeenSuspended());
+        exoPlayer,
+        videoPlayerEvents,
+        playerHasBeenSuspended(),
+        surfaceProducerHandlesCropAndRotation);
   }
 
   @RestrictTo(RestrictTo.Scope.LIBRARY)