+++ packages/video_player/video_player_android/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 2.8.1
+* Changes plugin to use `TextureRegistry.SurfaceProducer.handlesCropAndRotation` to detect
+  whether or not the video player rotation needs to be corrected.
+  
 ## 2.8.0
 
 * Adds support for platform views as an optional way of displaying a video.